requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
findspark==2.0.1
pyspark-kafka==0.10.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
tabulate==0.9.0
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0 